requirements.txt
discord.py>=2.3.2
Pillow>=10.4.0
Flask>=3.0.2
python-dotenv>=1.0.1
requests>=2.31.0
